Bosses

+++ src/main/java/dungeonmania/entities/collectableEntities/Sword.java
@@ -7,8 +7,8 @@ public class Sword extends Weapon {
 
     public Sword(String type, Position position) {
         super(type, position);
-        setDurability(JSONConfig.getConfig("sword_durability"));
-        setDamageValue(JSONConfig.getConfig("sword_attack"));
+        setDurability((int) JSONConfig.getConfig("sword_durability"));
+        setDamageValue((int) JSONConfig.getConfig("sword_attack"));
         setDefence(0);
 
     }
